requirements/test-3rd-party.txt
# SHA1:e410a404e60b5b015a51bf4636bfc2cf16813852
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
-r release-3rd-party.txt
alabaster==1.0.0
    # via sphinx
astroid==3.3.5
    # via pylint
babel==2.16.0
    # via sphinx
black==24.10.0
    # via -r requirements\test-3rd-party.in
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via black
colorama==0.4.6
    # via
    #   click
    #   pylint
    #   pytest
    #   sphinx
coverage[toml]==7.6.4
    # via pytest-cov
dill==0.3.9
    # via pylint
docutils==0.21.2
    # via sphinx
idna==3.10
    # via requests
imagesize==1.4.1
    # via sphinx
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via pylint
jinja2==3.1.4
    # via
    #   pytest-html
    #   sphinx
markupsafe==3.0.2
    # via jinja2
mccabe==0.7.0
    # via pylint
mypy==1.13.0
    # via -r requirements\test-3rd-party.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
numpydoc==1.8.0
    # via -r requirements\test-3rd-party.in
pandas-stubs==2.2.3.241009
    # via -r requirements\test-3rd-party.in
pathspec==0.12.1
    # via black
platformdirs==4.3.6
    # via
    #   black
    #   pylint
pluggy==1.5.0
    # via pytest
pygments==2.18.0
    # via sphinx
pylint==3.3.1
    # via -r requirements\test-3rd-party.in
pytest==8.3.3
    # via
    #   -r requirements\test-3rd-party.in
    #   pytest-cov
    #   pytest-html
    #   pytest-metadata
pytest-cov==6.0.0
    # via -r requirements\test-3rd-party.in
pytest-html==4.1.1
    # via -r requirements\test-3rd-party.in
pytest-metadata==3.1.1
    # via pytest-html
requests==2.32.3
    # via sphinx
snowballstemmer==2.2.0
    # via sphinx
sphinx==8.1.3
    # via
    #   -r requirements\test-3rd-party.in
    #   numpydoc
sphinxcontrib-applehelp==2.0.0
    # via sphinx
sphinxcontrib-devhelp==2.0.0
    # via sphinx
sphinxcontrib-htmlhelp==2.1.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==2.0.0
    # via sphinx
sphinxcontrib-serializinghtml==2.0.0
    # via sphinx
tabulate==0.9.0
    # via numpydoc
tomlkit==0.13.2
    # via pylint
types-pytz==2024.2.0.20241003
    # via pandas-stubs
typing-extensions==4.12.2
    # via mypy
urllib3==2.2.3
    # via requests
wheel==0.44.0
    # via -r requirements\test-3rd-party.in
